+++ crates/spaces/src/builtins/info.rs
@@ -1,11 +1,9 @@
-use crate::{rules, singleton};
+use crate::singleton;
 use anyhow::Context;
 use anyhow_source_location::{format_context, format_error};
 use starlark::environment::GlobalsBuilder;
 use starlark::values::none::NoneType;
-use starlark::values::{Heap, Value};
 use starstd::{Arg, Function};
-use std::collections::HashMap;
 
 pub const FUNCTIONS: &[Function] = &[
     Function {
@@ -41,121 +39,6 @@ pub const FUNCTIONS: &[Function] = &[
         description: "returns the path to the spaces store (typically $HOME/.spaces/store)",
         return_type: "str",
         args: &[],
-        example: None,
-    },
-    Function {
-        name: "get_path_to_log_file",
-        description: "returns the relative workspace path to the log file for the target",
-        return_type: "str",
-        args: &[            
-            Arg {
-                name: "name",
-                description: "The name of the target to get the log file",
-                dict: &[],
-            },
-        ],
-        example: None,
-    },
-    Function {
-        name: "get_path_to_workspace_member",
-        description: "returns a string to the workspace member matching the specified requirement (error if not found)",
-        return_type: "str",
-        args: &[            
-            Arg {
-                name: "member",
-                description: "The requirements for the member",
-                dict: &[
-                    ("url:str", "The url of the member"),
-                    ("required:dict", "{'Revision': <git/sha256 hash>}|{'SemVer': <semver requirement>}"),
-                ],
-            },
-        ],
-        example: None,
-    },
-    Function {
-        name: "is_path_to_workspace_member_available",
-        description: "returns true if the workspace satisfies the requirments",
-        return_type: "bool",
-        args: &[            
-            Arg {
-                name: "member",
-                description: "The requirements for the member",
-                dict: &[
-                    ("url:str", "The url of the member"),
-                    ("required:dict", "{'Revision': <git/sha256 hash>}|{'SemVer': <semver requirement>}"),
-                ],
-            },
-        ],
-        example: None,
-    },
-    Function {
-        name: "get_absolute_path_to_workspace",
-        description: "returns the absolute path to the workspace",
-        return_type: "str",
-        args: &[],
-        example: None,
-    },
-    Function {
-        name: "get_path_to_checkout",
-        description: "returns the path where the current script is located in the workspace",
-        return_type: "str",
-        args: &[],
-        example: None,
-    },
-    Function {
-        name: "get_env_var",
-        description: "returns the path where the current script is located in the workspace",
-        return_type: "str",
-        args: &[
-            Arg {
-                name: "var",
-                description: "The name of the environment variable",
-                dict: &[],
-            },
-        ],
-        example: None,
-    },
-    Function {
-        name: "get_path_to_build_checkout",
-        description: "returns the path to the workspace build folder for the current script",
-        return_type: "str",
-        args: &[],
-        example: None,
-    },
-    Function {
-        name: "get_path_to_build_archive",
-        description: "returns the path to where run.create_archive() creates the output archive",
-        return_type: "str",
-        args: &[
-            Arg {
-                name: "rule_name",
-                description: "The name of the rule used to create the archive",
-                dict: &[],
-            },
-            Arg {
-                name: "archive",
-                description: "The archive info used to create the archive",
-                dict: &[],
-            },
-        ],
-        example: None,
-    },
-    Function {
-        name: "get_build_archive_info",
-        description: "returns the path to where run.create_archive() creates the sha256 txt file",
-        return_type: "dict['archive_path': str, 'sha256_path': str]",
-        args: &[
-            Arg {
-                name: "rule_name",
-                description: "The name of the rule used to create the archive",
-                dict: &[],
-            },
-            Arg {
-                name: "archive",
-                description: "The archive info used to create the archive",
-                dict: &[],
-            },
-        ],
         example: None,
     },
     Function {
@@ -216,39 +99,6 @@ pub const FUNCTIONS: &[Function] = &[
 
 #[starlark_module]
 pub fn globals(builder: &mut GlobalsBuilder) {
-    // remove and replace with get_path_to_store()
-    fn store_path() -> anyhow::Result<String> {
-        let workspace_arc =
-            singleton::get_workspace().context(format_error!("No active workspace found"))?;
-        let workspace = workspace_arc.read();
-        Ok(workspace.get_store_path().to_string())
-    }
-
-    // remove and replace with get_absolute_path_to_workspace()
-    fn absolute_workspace_path() -> anyhow::Result<String> {
-        let workspace_arc =
-            singleton::get_workspace().context(format_error!("No active workspace found"))?;
-        let absolute_path = workspace_arc.read().get_absolute_path();
-        Ok(absolute_path.to_string())
-    }
-
-    // remove and replace with get_platform_name()
-    fn platform_name() -> anyhow::Result<String> {
-        platform::Platform::get_platform()
-            .map(|p| p.to_string())
-            .ok_or(anyhow::anyhow!("Failed to get platform name"))
-    }
-
-    // remove and replace with get_path_to_checkout()
-    fn checkout_path() -> anyhow::Result<String> {
-        rules::get_checkout_path().map(|p| p.to_string())
-    }
-
-    // remove and replace with get_path_to_checkout()
-    fn current_workspace_path() -> anyhow::Result<String> {
-        rules::get_checkout_path().map(|p| p.to_string())
-    }
-
     fn get_platform_name() -> anyhow::Result<String> {
         platform::Platform::get_platform()
             .map(|p| p.to_string())
@@ -260,27 +110,6 @@ pub fn globals(builder: &mut GlobalsBuilder) {
             .into_iter()
             .map(|p| p.to_string())
             .collect())
-    }
-
-    fn is_workspace_reproducible() -> anyhow::Result<bool> {
-        let workspace_arc =
-            singleton::get_workspace().context(format_error!("No active workspace found"))?;
-        let workspace = workspace_arc.read();
-        Ok(workspace.is_reproducible())
-    }
-
-    fn get_workspace_digest() -> anyhow::Result<String> {
-        let workspace_arc =
-            singleton::get_workspace().context(format_error!("No active workspace found"))?;
-        let workspace = workspace_arc.read();
-        Ok(workspace.digest.clone().to_string())
-    }
-
-    fn get_workspace_short_digest() -> anyhow::Result<String> {
-        let workspace_arc =
-            singleton::get_workspace().context(format_error!("No active workspace found"))?;
-        let workspace = workspace_arc.read();
-        Ok(workspace.get_short_digest().to_string())
     }
 
     fn is_ci() -> anyhow::Result<bool> {
@@ -311,106 +140,8 @@ pub fn globals(builder: &mut GlobalsBuilder) {
         Err(format_error!("Info Aborting: {}", message))
     }
 
-    fn is_env_var_set(var_name: &str) -> anyhow::Result<bool> {
-        if var_name == "PATH" {
-            return Ok(true);
-        }
-        let workspace_arc =
-            singleton::get_workspace().context(format_error!("No active workspace found"))?;
-        let workspace = workspace_arc.read();
-        Ok(workspace.get_env().vars.contains_key(var_name))
-    }
-
-    fn get_env_var(var_name: &str) -> anyhow::Result<String> {
-        let workspace_arc =
-            singleton::get_workspace().context(format_error!("No active workspace found"))?;
-        let workspace = workspace_arc.read();
-        let env = workspace.get_env();
-        if var_name == "PATH" {
-            return Ok(env.get_path().to_string());
-        }
-
-        if let Some(value) = env.vars.get(var_name) {
-            return Ok(value.clone().to_string());
-        }
-
-        Err(format_error!(
-            "{var_name} is not set in the workspace environment"
-        ))
-    }
-
-    fn set_env(
-        #[starlark(require = named)] env: starlark::values::Value,
-    ) -> anyhow::Result<NoneType> {
-        let workspace_arc =
-            singleton::get_workspace().context(format_error!("No active workspace found"))?;
-
-        let mut workspace = workspace_arc.write();
-        let env = serde_json::from_value(env.to_json_value()?)
-            .context(format_context!("Failed to parse archive arguments"))?;
-
-        workspace.set_env(env);
-
-        Ok(NoneType)
-    }
-
-    fn set_locks(
-        #[starlark(require = named)] locks: starlark::values::Value,
-    ) -> anyhow::Result<NoneType> {
-        let locks = serde_json::from_value(locks.to_json_value()?)
-            .context(format_context!("Failed to parse archive arguments"))?;
-
-        let workspace_arc = singleton::get_workspace()
-            .context(format_error!("Internal Error: No active workspace found"))?;
-        let mut workspace = workspace_arc.write();
-
-        workspace.locks = locks;
-
-        Ok(NoneType)
-    }
-
-    fn get_path_to_workspace_member(
-        #[starlark(require = named)] member: starlark::values::Value,
-    )-> anyhow::Result<String>{
-        let workspace_arc = singleton::get_workspace()
-        .context(format_error!("Internal Error: No active workspace found"))?;
-        let member_requirement_json = member.to_json_value()?;
-        let member_requirement: ws::MemberRequirement = serde_json::from_value(member_requirement_json.clone())
-            .context(format_context!("bad options for workspace member"))?;
-
-        let path = workspace_arc.read().settings.get_path_to_member(&member_requirement);
-        match path {
-            Some(p) => Ok(p.to_string()),
-            None => Err(format_error!("`{}` not found in workspace matching {:?}", member_requirement.url, member_requirement.required)),
-        }
-    }
-
-    fn is_path_to_workspace_member_available(
-        #[starlark(require = named)] member: starlark::values::Value,
-    )-> anyhow::Result<bool>{
-        let workspace_arc = singleton::get_workspace()
-        .context(format_error!("Internal Error: No active workspace found"))?;
-        let member_requirement_json = member.to_json_value()?;
-        let member_requirement: ws::MemberRequirement = serde_json::from_value(member_requirement_json.clone())
-            .context(format_context!("bad options for workspace member"))?;
-
-        let path = workspace_arc.read().settings.get_path_to_member(&member_requirement);
-        match path {
-            Some(_) => Ok(true),
-            None => Ok(false),
-        }
-    }
-
     fn get_cpu_count() -> anyhow::Result<i64> {
         Ok(num_cpus::get() as i64)
-    }
-
-    fn get_path_to_log_file(target: &str) -> anyhow::Result<String> {
-        let workspace_arc =
-            singleton::get_workspace().context(format_error!("No active workspace found"))?;
-        let workspace = workspace_arc.read();
-        let rule_name = rules::get_sanitized_rule_name(target.into());
-        Ok(workspace.get_log_file(rule_name.as_ref()).to_string())
     }
 
     fn get_path_to_store() -> anyhow::Result<String> {
@@ -420,80 +151,11 @@ pub fn globals(builder: &mut GlobalsBuilder) {
         Ok(workspace.get_store_path().to_string())
     }
 
-    fn get_absolute_path_to_workspace() -> anyhow::Result<String> {
-        let workspace_arc =
-            singleton::get_workspace().context(format_error!("No active workspace found"))?;
-        let workspace = workspace_arc.read();
-        Ok(workspace.absolute_path.clone().to_string())
-    }
-
-    fn get_path_to_checkout() -> anyhow::Result<String> {
-        rules::get_checkout_path().map(|p| p.to_string())
-    }
-
-    fn get_path_to_build_checkout(
-        #[starlark(require = named)] rule_name: &str,
-    ) -> anyhow::Result<String> {
-        rules::get_path_to_build_checkout(rule_name.into()).map(|p| p.to_string())
-    }
-
-    fn get_path_to_build_archive(
-        #[starlark(require = named)] rule_name: &str,
-        #[starlark(require = named)] archive: starlark::values::Value,
-    ) -> anyhow::Result<String> {
-        let create_archive: easy_archiver::CreateArchive =
-            serde_json::from_value(archive.to_json_value()?)
-                .context(format_context!("bad options for archive"))?;
-
-        let sanitized_rule_name = rules::get_sanitized_rule_name(rule_name.into());
-
-        Ok(format!(
-            "build/{sanitized_rule_name}/{}",
-            create_archive.get_output_file()
-        ))
-    }
-
     fn get_path_to_spaces_tools() -> anyhow::Result<String> {
         let workspace_arc =
             singleton::get_workspace().context(format_error!("No active workspace found"))?;
         let workspace = workspace_arc.read();
         Ok(workspace.get_spaces_tools_path().to_string())
-    }
-
-    fn get_build_archive_info<'v>(
-        #[starlark(require = named)] rule_name: &str,
-        #[starlark(require = named)] archive: starlark::values::Value,
-        heap: &'v Heap,
-    ) -> anyhow::Result<Value<'v>> {
-        let create_archive: easy_archiver::CreateArchive =
-            serde_json::from_value(archive.to_json_value()?)
-                .context(format_context!("bad options for archive"))?;
-
-        let create_archive_output = create_archive.get_output_file();
-        let output_path = std::path::Path::new(create_archive_output.as_str());
-        let output_sha_suffix = output_path.with_extension("").with_extension("sha256.txt");
-
-        let sanitized_rule_name = rules::get_sanitized_rule_name(rule_name.into());
-
-        let mut output = HashMap::new();
-        let rule_output_path = format!("build/{sanitized_rule_name}");
-
-        output.insert(
-            "archive_path".to_string(),
-            format!("{rule_output_path}/{create_archive_output}",),
-        );
-        output.insert(
-            "sha256_path".to_string(),
-            format!("{rule_output_path}/{}", output_sha_suffix.to_string_lossy()),
-        );
-
-        let json_value = serde_json::to_value(&output)
-            .context(format_context!("Failed to convert Result to JSON"))?;
-
-        // Convert the JSON value to a Starlark value
-        let alloc_value = heap.alloc(json_value);
-
-        Ok(alloc_value)
     }
 
     fn set_minimum_version(version: &str) -> anyhow::Result<NoneType> {